Moved venues route to its own file in /routes
